KL-176/refactor: separate location state initialization into custom hook
